Refactor code to move Home component logic to separate file and update router paths to include '/recipe' prefix
