+++ app/components/Header.js
@@ -37,7 +37,7 @@ const Header = () => {
               </svg>
             </span>
           </Link>
-          <Link href="/recipes">
+          <Link href="/recipe">
             <span className="hover:text-gray-500 cursor-pointer">Recipes</span>
           </Link>
           <Link href="/favourites">
